+++ BuildResidentialHPXML/measure.rb
@@ -1002,8 +1002,8 @@ class BuildResidentialHPXML < OpenStudio::Measure::ModelMeasure
     compressor_type_choices << HPXML::HVACCompressorTypeTwoStage
     compressor_type_choices << HPXML::HVACCompressorTypeVariableSpeed
 
-    attached_heating_system_fuel_choices = OpenStudio::StringVector.new
-    attached_heating_system_fuel_choices << HPXML::FuelTypeElectricity
+    integrated_heating_system_fuel_choices = OpenStudio::StringVector.new
+    integrated_heating_system_fuel_choices << HPXML::FuelTypeElectricity
 
     arg = OpenStudio::Measure::OSArgument::makeChoiceArgument('heating_system_type', heating_system_type_choices, true)
     arg.setDisplayName('Heating System: Type')
@@ -1103,24 +1103,24 @@ class BuildResidentialHPXML < OpenStudio::Measure::ModelMeasure
     arg.setUnits('Frac')
     args << arg
 
-    arg = OpenStudio::Measure::OSArgument::makeChoiceArgument('cooling_system_attached_heating_system_fuel', attached_heating_system_fuel_choices, false)
+    arg = OpenStudio::Measure::OSArgument::makeChoiceArgument('cooling_system_integrated_heating_system_fuel', integrated_heating_system_fuel_choices, false)
     arg.setDisplayName('Cooling System: Attached Heating System Fuel Type')
     arg.setDescription("The fuel type of the heating system attached to cooling system. Only used for #{HPXML::HVACTypePTAC} and #{HPXML::HVACTypeRoomAirConditioner}.")
     args << arg
 
-    arg = OpenStudio::Measure::OSArgument::makeDoubleArgument('cooling_system_attached_heating_system_efficiency', false)
+    arg = OpenStudio::Measure::OSArgument::makeDoubleArgument('cooling_system_integrated_heating_system_efficiency', false)
     arg.setDisplayName('Cooling System: Attached Heating System Efficiency')
     arg.setUnits('Frac')
     arg.setDescription("The rated heating efficiency value of the heating system attached to cooling system. Only used for #{HPXML::HVACTypePTAC} and #{HPXML::HVACTypeRoomAirConditioner}.")
     args << arg
 
-    arg = OpenStudio::Measure::OSArgument::makeDoubleArgument('cooling_system_attached_heating_system_capacity', false)
+    arg = OpenStudio::Measure::OSArgument::makeDoubleArgument('cooling_system_integrated_heating_system_capacity', false)
     arg.setDisplayName('Cooling System: Attached Heating System Heating Capacity')
     arg.setDescription("The output heating capacity of the heating system attached to cooling system. If not provided, size the capacity based on ACCA Manual J/S. Only used for #{HPXML::HVACTypePTAC} and #{HPXML::HVACTypeRoomAirConditioner}.")
     arg.setUnits('Btu/hr')
     args << arg
 
-    arg = OpenStudio::Measure::OSArgument::makeDoubleArgument('cooling_system_attached_heating_system_fraction_heat_load_served', false)
+    arg = OpenStudio::Measure::OSArgument::makeDoubleArgument('cooling_system_integrated_heating_system_fraction_heat_load_served', false)
     arg.setDisplayName('Cooling System: Attached Heating System Fraction Heat Load Served')
     arg.setDescription("The heating load served by the heating system attached to cooling system. Only used for #{HPXML::HVACTypePTAC} and #{HPXML::HVACTypeRoomAirConditioner}.")
     arg.setUnits('Frac')
@@ -4586,20 +4586,20 @@ class HPXMLFile
     end
 
     if [HPXML::HVACTypePTAC, HPXML::HVACTypeRoomAirConditioner].include?(cooling_system_type)
-      if args[:cooling_system_attached_heating_system_fuel].is_initialized
-        attached_heating_system_fuel = args[:cooling_system_attached_heating_system_fuel].get
+      if args[:cooling_system_integrated_heating_system_fuel].is_initialized
+        integrated_heating_system_fuel = args[:cooling_system_integrated_heating_system_fuel].get
       end
 
-      if args[:cooling_system_attached_heating_system_fraction_heat_load_served].is_initialized
-        attached_heating_system_fraction_heat_load_served = args[:cooling_system_attached_heating_system_fraction_heat_load_served].get
+      if args[:cooling_system_integrated_heating_system_fraction_heat_load_served].is_initialized
+        integrated_heating_system_fraction_heat_load_served = args[:cooling_system_integrated_heating_system_fraction_heat_load_served].get
       end
 
-      if args[:cooling_system_attached_heating_system_capacity].is_initialized
-        attached_heating_system_capacity = args[:cooling_system_attached_heating_system_capacity].get
+      if args[:cooling_system_integrated_heating_system_capacity].is_initialized
+        integrated_heating_system_capacity = args[:cooling_system_integrated_heating_system_capacity].get
       end
 
-      if args[:cooling_system_attached_heating_system_efficiency].is_initialized
-        attached_heating_system_efficiency = args[:cooling_system_attached_heating_system_efficiency].get
+      if args[:cooling_system_integrated_heating_system_efficiency].is_initialized
+        integrated_heating_system_efficiency = args[:cooling_system_integrated_heating_system_efficiency].get
       end
     end
 
@@ -4617,10 +4617,10 @@ class HPXMLFile
                               airflow_defect_ratio: airflow_defect_ratio,
                               charge_defect_ratio: charge_defect_ratio,
                               primary_system: true,
-                              attached_heating_system_fuel: attached_heating_system_fuel,
-                              attached_heating_system_capacity: attached_heating_system_capacity,
-                              attached_heating_system_efficiency: attached_heating_system_efficiency,
-                              attached_heating_system_fraction_heat_load_served: attached_heating_system_fraction_heat_load_served)
+                              integrated_heating_system_fuel: integrated_heating_system_fuel,
+                              integrated_heating_system_capacity: integrated_heating_system_capacity,
+                              integrated_heating_system_efficiency: integrated_heating_system_efficiency,
+                              integrated_heating_system_fraction_heat_load_served: integrated_heating_system_fraction_heat_load_served)
   end
 
   def self.set_heat_pumps(hpxml, args)