+++ HPXMLtoOpenStudio/tests/test_hvac.rb
@@ -138,9 +138,9 @@ class HPXMLtoOpenStudioHVACTest < MiniTest::Test
     cop = UnitConversions.convert(ceer, 'Btu/hr', 'W') # Expected value
     cool_capacity = UnitConversions.convert(cooling_system.cooling_capacity, 'Btu/hr', 'W')
 
-    heat_efficiency = cooling_system.attached_heating_system_efficiency
+    heat_efficiency = cooling_system.integrated_heating_system_efficiency
     heat_efficiency = 1.0 if heat_efficiency.nil?
-    heat_capacity = UnitConversions.convert(cooling_system.attached_heating_system_capacity, 'Btu/hr', 'W')
+    heat_capacity = UnitConversions.convert(cooling_system.integrated_heating_system_capacity, 'Btu/hr', 'W')
 
     # Check cooling coil
     assert_equal(1, model.getAirLoopHVACUnitarySystems.size)
@@ -188,9 +188,9 @@ class HPXMLtoOpenStudioHVACTest < MiniTest::Test
     cop = UnitConversions.convert(ceer, 'Btu/hr', 'W') # Expected value
     cool_capacity = UnitConversions.convert(cooling_system.cooling_capacity, 'Btu/hr', 'W')
 
-    heat_efficiency = cooling_system.attached_heating_system_efficiency
+    heat_efficiency = cooling_system.integrated_heating_system_efficiency
     heat_efficiency = 1.0 if heat_efficiency.nil?
-    heat_capacity = UnitConversions.convert(cooling_system.attached_heating_system_capacity, 'Btu/hr', 'W')
+    heat_capacity = UnitConversions.convert(cooling_system.integrated_heating_system_capacity, 'Btu/hr', 'W')
 
     # Check cooling coil
     assert_equal(1, model.getAirLoopHVACUnitarySystems.size)