+++ tasks.rb
@@ -266,9 +266,9 @@ def create_hpxmls
     'base-hvac-autosize-pthp-sizing-methodology-acca.xml' => 'base-hvac-pthp.xml',
     'base-hvac-autosize-pthp-sizing-methodology-hers.xml' => 'base-hvac-pthp.xml',
     'base-hvac-autosize-pthp-sizing-methodology-maxload.xml' => 'base-hvac-pthp.xml',
-    'base-hvac-autosize-room-heat-pump-sizing-methodology-acca.xml' => 'base-hvac-room-heat-pump.xml',
-    'base-hvac-autosize-room-heat-pump-sizing-methodology-hers.xml' => 'base-hvac-room-heat-pump.xml',
-    'base-hvac-autosize-room-heat-pump-sizing-methodology-maxload.xml' => 'base-hvac-room-heat-pump.xml',
+    'base-hvac-autosize-room-ac-with-reverse-cycle-sizing-methodology-acca.xml' => 'base-hvac-room-ac-with-reverse-cycle.xml',
+    'base-hvac-autosize-room-ac-with-reverse-cycle-sizing-methodology-hers.xml' => 'base-hvac-room-ac-with-reverse-cycle.xml',
+    'base-hvac-autosize-room-ac-with-reverse-cycle-sizing-methodology-maxload.xml' => 'base-hvac-room-ac-with-reverse-cycle.xml',
     'base-hvac-autosize-room-ac-with-heating.xml' => 'base-hvac-room-ac-with-heating.xml',
     'base-hvac-autosize-room-ac-only.xml' => 'base-hvac-room-ac-only.xml',
     'base-hvac-autosize-stove-oil-only.xml' => 'base-hvac-stove-oil-only.xml',
@@ -344,7 +344,7 @@ def create_hpxmls
     'base-hvac-room-ac-only-33percent.xml' => 'base-hvac-room-ac-only.xml',
     'base-hvac-room-ac-only-ceer.xml' => 'base-hvac-room-ac-only.xml',
     'base-hvac-room-ac-with-heating.xml' => 'base-hvac-room-ac-only.xml',
-    'base-hvac-room-heat-pump.xml' => 'base-hvac-pthp.xml',
+    'base-hvac-room-ac-with-reverse-cycle.xml' => 'base-hvac-pthp.xml',
     'base-hvac-room-ac-only-detailed-setpoints.xml' => 'base-hvac-room-ac-only.xml',
     'base-hvac-seasons.xml' => 'base.xml',
     'base-hvac-setpoints.xml' => 'base.xml',
@@ -1748,7 +1748,7 @@ def set_measure_argument_values(hpxml_file, args, sch_args, orig_parent)
     args.delete('heating_system_2_heating_capacity')
     args.delete('cooling_system_cooling_capacity')
     args.delete('heat_pump_heating_capacity')
-    args.delete('cooling_system_attached_heating_system_capacity')
+    args.delete('cooling_system_integrated_heating_system_capacity')
     if hpxml_file.include? 'sizing-methodology-hers'
       args['heat_pump_sizing_methodology'] = HPXML::HeatPumpSizingHERS
     elsif hpxml_file.include? 'sizing-methodology-maxload'
@@ -1997,10 +1997,10 @@ def set_measure_argument_values(hpxml_file, args, sch_args, orig_parent)
     args['cooling_system_cooling_capacity'] = 8000.0
   elsif ['base-hvac-room-ac-with-heating.xml',
          'base-hvac-ptac-with-heating.xml'].include? hpxml_file
-    args['cooling_system_attached_heating_system_efficiency'] = 1.0
-    args['cooling_system_attached_heating_system_capacity'] = 36000.0
-    args['cooling_system_attached_heating_system_fraction_heat_load_served'] = 1.0
-    args['cooling_system_attached_heating_system_fuel'] = HPXML::FuelTypeElectricity
+    args['cooling_system_integrated_heating_system_efficiency'] = 1.0
+    args['cooling_system_integrated_heating_system_capacity'] = 36000.0
+    args['cooling_system_integrated_heating_system_fraction_heat_load_served'] = 1.0
+    args['cooling_system_integrated_heating_system_fuel'] = HPXML::FuelTypeElectricity
   elsif ['base-hvac-setpoints.xml'].include? hpxml_file
     args['hvac_control_heating_weekday_setpoint'] = 60
     args['hvac_control_heating_weekend_setpoint'] = 60
@@ -2027,7 +2027,7 @@ def set_measure_argument_values(hpxml_file, args, sch_args, orig_parent)
     args['heat_pump_type'] = HPXML::HVACTypeHeatPumpPTHP
     args['heat_pump_cooling_efficiency'] = 11.4
     args['heat_pump_cooling_sensible_heat_fraction'] = 0.65
-  elsif ['base-hvac-room-heat-pump.xml'].include? hpxml_file
+  elsif ['base-hvac-room-ac-with-reverse-cycle.xml'].include? hpxml_file
     args['heat_pump_type'] = HPXML::HVACTypeHeatPumpRoom
   elsif ['base-hvac-ptac.xml'].include? hpxml_file
     args['heating_system_type'] = 'none'